added new module: phx_gui
This is a ros python node which sends messages to the osc network. the counterpart 'base_station' receives this via osc and can plot the data. It is also possible to send parameters from the gui via osc directly into the ros network.

python osc and pyQT4 is needed for this software to run.